# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PayGate Payment Platform API
  version: 1.0.0
  description: |
    API contract for the current MVP implementation of the PayGate multi-tenant
    payment middleware.

    Coverage in this spec matches the endpoints that are already live in the
    backend codebase today:
    - dashboard auth and MFA
    - store management
    - store API token management
    - store-facing transaction API
    - Midtrans inbound webhook
    - webhook delivery inspection and resend
    - dashboard operational alert destinations
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/dashboard/auth/mfa/recovery/regenerate:
    post:
      tags: [Dashboard Auth]
      summary: Regenerate MFA recovery codes
      operationId: regenerateRecoveryCodes
      security:
        - DashboardBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFACodeRequest"
      responses:
        "200":
          description: New recovery codes generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RecoveryCodesResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/MFAInvalidCodeError"
        "409":
          $ref: "#/components/responses/MFASetupRequiredError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    DashboardBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    StoreBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Store API Token
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid payload.
              request_id: req_abc123
              details: {}
...
    MFASetupRequiredError:
      description: MFA has not been enabled yet
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: MFA_SETUP_REQUIRED
              message: MFA is not enabled yet.
              request_id: req_abc123
              details: {}
...
    MFAInvalidCodeError:
      description: MFA code or recovery code is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: MFA_CODE_INVALID
              message: Invalid MFA or recovery code.
              request_id: req_abc123
              details: {}
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: Internal server error.
              request_id: req_abc123
              details: {}
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data: {}
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message, request_id, details]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
          additionalProperties: true
...
    MFACodeRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: TOTP code or recovery code depending on endpoint and state
...
    RecoveryCodesResponse:
      type: object
      required: [recovery_codes]
      properties:
        recovery_codes:
          type: array
          items:
            type: string
...